pyproject.toml
[project]
name = "medhackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "conformer",
    "gdown>=5.2.0",
    "google-genai>=1.10.0",
    "gradio>=5.25.0",
    "kagglehub>=0.3.11",
    "notebook>=7.4.0",
    "openai-whisper",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "streamlit>=1.44.1",
]

[tool.uv.sources]
openai-whisper = { git = "https://github.com/openai/whisper.git" }
conformer = { git = "https://github.com/sooftware/conformer.git" }
